Use JSON in the log.

+++ src/Logger.php
@@ -26,9 +26,18 @@ class Logger
     public static function log(string $message): void
     {
         $path             = plugin_dir_path(__DIR__);
-        $message          = wp_json_encode($message) . "\n";
-        $destination_file = $path . 'c2w.log';
-        error_log(htmlspecialchars($message), 3, $destination_file); // phpcs:ignore
+        $logFile          = $path . 'c2w.log.json';
+
+        if (file_exists($logFile)) {
+            $logContent       = json_decode(file_get_contents($logFile));
+        } else {
+            $logContent = [];
+        }
+
+        $message          = htmlspecialchars($message);
+        $messageItem      = ["message" => "$message"];
+        $logContent[]     = $messageItem;
+        file_put_contents($logFile, wp_json_encode($logContent));
     }
 
     /**
@@ -39,7 +48,9 @@ class Logger
     public static function clear(): void
     {
         $path             = plugin_dir_path(__DIR__);
-        $destination_file = $path . 'c2w.log';
-        unlink($destination_file);
+        $destination_file = $path . 'c2w.log.json';
+        if (file_exists($destination_file)) {
+            unlink($destination_file);
+        }
     }
 }
